WIP: refactoring lib code
